requirements.txt
django==5.0
Pillow==10.1.0
SimConnect==0.4.26
celery==5.3.6
channels[daphne]~=4.0.0
channels_redis==4.1
coverage==7.3.4
crispy-bootstrap4==2023.1
django-bootstrap-dynamic-formsets==0.5.0
django-bootstrap4==23.2
django-countries==7.5.1
django-crispy-forms==2.1
django-formtools==2.5.1
django-guardian==2.4.0
django-phonenumber-field==7.2.0
django-qr-code==3.1.1
django-solo==2.1.0
django-storages[azure, google]
django-timezone-field==6.1.0
django-use-email-as-username==1.4.0
django-webpack-loader==3.0.0
djangorestframework-guardian==0.3.0
djangorestframework==3.14.0
drf-firebase-auth==1.0.0
drf-nested-routers==0.93.5
drf-yasg==1.21.7
fastkml==0.12
firebase-admin==4.5.3
flower==2.0.1
fpdf2==2.7.7
geopy==2.4.1
gpx-converter==2.1.0
gpxpy==1.6.2
hiredis==2.3.2
html2text==2020.1.16
iso8601==2.1.0
kubernetes==28.1.0
kubernetes-job==0.3.3
lxml==4.9.4
matplotlib==3.8.2
mysqlclient==2.2.1
numpy==1.26.2
nvector==0.7.7
#ogn-client==1.2.1
pandas==2.1.4
parameterized==0.9.0
pep8==1.7.1
phonenumbers==8.13.27
pulp==2.7.0
pyepsg==0.4.0
pykalman==0.9.5
pylatex==1.4.2
pymbtiles==0.5.0
pytest
pytest-django
pytest-cov
pyproj==3.6.1
python-redis-lock==4.0.0
pytz==2023.3
qrcode==7.4.2
redis[hiredis]==5.0.1
requests==2.31.0
scipy==1.11.4
#segno==1.6.0
twisted==23.10.0
tzlocal==5.2
utm==0.7.0
uvicorn[standard]==0.25.0
websocket-client==1.7.0
whitenoise==6.6.0
django-location-field==2.7.3
cartopy==0.22.0
google-cloud-logging==3.9.0
python-dateutil~=2.8.2
